+++ Application/src/gestionbrb/model/Produit.java
@@ -1,17 +1,15 @@
 package gestionbrb.model;
 
 
-import javafx.beans.property.FloatProperty;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleFloatProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 /**
- * <!-- begin-user-doc -->
- * <!--  end-user-doc  -->
- * @generated
+ * Produits que les clients commandent
  */
 
 public class Produit {
@@ -20,26 +18,26 @@ public class Produit {
 	private final StringProperty nomProduit;
 	private final IntegerProperty quantiteProduit;
 	private final StringProperty descriptionProduit;
-	private final FloatProperty prixProduit;
+	private final DoubleProperty prixProduit;
 	private final StringProperty type;
 	private final StringProperty ingredients;
 
-	public Produit(int idProduit, String nomProduit, int quantiteProduit, String descriptionProduit, float prixProduit, String type, String ingredients){
+	public Produit(int idProduit, String nomProduit, int quantiteProduit, String descriptionProduit, double prixProduit, String type, String ingredients){
 		this.idProduit = new SimpleIntegerProperty(idProduit);
 		this.nomProduit = new SimpleStringProperty(nomProduit);
 		this.quantiteProduit = new SimpleIntegerProperty(quantiteProduit);
 		this.descriptionProduit = new SimpleStringProperty(descriptionProduit);
-		this.prixProduit = new SimpleFloatProperty(prixProduit);
+		this.prixProduit = new SimpleDoubleProperty(prixProduit);
 		this.type = new SimpleStringProperty(type);
 		this.ingredients = new SimpleStringProperty(ingredients);
 	}
 	
-	public Produit(String nomProduit, float prixProduit, int quantiteProduit){
+	public Produit(String nomProduit, double prixProduit, int quantiteProduit){
 		this.idProduit = new SimpleIntegerProperty(0);
 		this.nomProduit = new SimpleStringProperty(nomProduit);
 		this.quantiteProduit = new SimpleIntegerProperty(quantiteProduit);
 		this.descriptionProduit = new SimpleStringProperty(null);
-		this.prixProduit = new SimpleFloatProperty(prixProduit);
+		this.prixProduit = new SimpleDoubleProperty(prixProduit);
 		this.type = new SimpleStringProperty(null);
 		this.ingredients = new SimpleStringProperty(null);
 	}
@@ -65,13 +63,13 @@ public class Produit {
 	public final void  setNomProduit(final String nomProduit) {
 		this.nomProduitProperty().set(nomProduit);
 	}
-	public final FloatProperty prixProduitProperty() {
+	public final DoubleProperty prixProduitProperty() {
 		return this.prixProduit;
 	}
-	public final float getPrixProduit() {
+	public final double getPrixProduit() {
 		return this.prixProduitProperty().get();
 	}
-	public final void  setPrixProduit(final float prixProduit) {
+	public final void  setPrixProduit(final double prixProduit) {
 		this.prixProduitProperty().set(prixProduit);
 	}
 	public final IntegerProperty quantiteProduitProperty() {
